+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## [0.2.7]
+
+### Added
+- The Websocket server url can now be provided into the SDK.
+- The `createWallet` method has been brought back along with
+its corresponding example `create_wallet.dart`.
+
 ## [0.2.6]
 
 ### Fixed